move custom converters to package-private/internal
