+++ integration-test/kotlin/spring-app/src/main/kotlin/com/qualityminds/lazyval/integration/boundary/persistence/DateToOrderDateConverter.kt
@@ -1,13 +1,13 @@
-package com.qualityminds.lazyval.integration.boundary.persistence.mongodb
+package com.qualityminds.lazyval.integration.boundary.persistence
 
 import com.qualityminds.lazyval.integration.domain.OrderDate
 import org.springframework.core.convert.converter.Converter
 import org.springframework.data.convert.ReadingConverter
 import java.time.ZoneOffset
-import java.util.Date
+import java.util.*
 
 @ReadingConverter
-class DateToOrderDateConverter : Converter<Date, OrderDate> {
+internal class DateToOrderDateConverter : Converter<Date, OrderDate> {
     override fun convert(source: Date): OrderDate =
         OrderDate(source.toInstant().atZone(ZoneOffset.UTC).toLocalDate())
 }
