+++ integration-test/kotlin/spring-app/src/main/kotlin/com/qualityminds/lazyval/integration/boundary/persistence/OrderDateToDateConverter.kt
@@ -1,13 +1,13 @@
-package com.qualityminds.lazyval.integration.boundary.persistence.mongodb
+package com.qualityminds.lazyval.integration.boundary.persistence
 
 import com.qualityminds.lazyval.integration.domain.OrderDate
 import org.springframework.core.convert.converter.Converter
 import org.springframework.data.convert.WritingConverter
 import java.time.ZoneOffset
-import java.util.Date
+import java.util.*
 
 @WritingConverter
-class OrderDateToDateConverter : Converter<OrderDate, Date> {
+internal class OrderDateToDateConverter : Converter<OrderDate, Date> {
     override fun convert(source: OrderDate): Date =
         Date.from(source.value.atStartOfDay(ZoneOffset.UTC).toInstant())
 }
